+++ CHANGELOG.md
@@ -17,6 +17,8 @@ n
 - Added new benchmark functions `make_benchmark_collection_2d_var_plots` and `make_benchmark_collection_3d_var_plots` which can be used with any GEOS-Chem output collection
 - Added 1-month benchmark comparison plot options for `Budget`, `UVFlux`, and `StateMet` collections (2D and 3D vars separately) which are off by default
 - Added `export MPLBACKEND=agg` to `gcpy/benchmark/modules/benchmark_slurm.sh` to request a non-interactive MatPlotLib backend
+- Added `gcpy/kpp/kpp_tolerances.py` to plot sorted results from KPP-Standalone tolerance loop output
+- Added routine `kppsa_read_tolerance_loop` in `gcpy/kpp/kppsa_utils.py`
 
 ### Changed
 - Updated `gcpy_environment_py313.yml` to use `esmf==8.8.1` and `esmpy==8.8.1` to fix package inconsistency issues